requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
bitarray==2.7.3
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cytoolz==0.12.1
ed25519==1.5
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-routes==0.1.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.2
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.22.4
py-ed25519-bindings==1.0.2
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyNaCl==1.5.0
pyotp==2.8.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pywin32==306
qrcode==7.4.2
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
six==1.16.0
SQLAlchemy==2.0.12
toolz==0.12.0
typing_extensions==4.5.0
urllib3==2.0.2
web3==6.3.0
websockets==11.0.2
Werkzeug==2.3.3
yarl==1.9.2
